mp4e: add codec info extraction and CLI tool
Add ExtractCodecInfo to the mp4e package, which parses MP4/fMP4 init
segments and returns per-track CodecInfo structs containing the RFC 6381
codec parameter string, profile/level IDC values, channel count, and
EC-3/Atmos-specific fields (channel map, JOC, complexity index).

Add ProfileName and LevelName lookup functions that convert profile/level
IDC integers to human-readable strings for AVC, HEVC, and AAC.

Add a cmd/codec-info CLI that runs ExtractCodecInfo on a file and prints
the result as JSON.

Upgrade github.com/Eyevinn/mp4ff from v0.49.0 to v0.51.0.

+++ mp4e/mp4.go
@@ -213,7 +213,7 @@ func ValidateFmp4(reader io.Reader) (file *mp4.File, info *Mp4Info, err error) {
 					info.SampleDurationMax = dur
 				}
 				dts := sample.DecodeTime
-				pts := sample.PresentationTime()
+				pts := uint64(sample.PresentationTime())
 
 				// Build list of samples
 				segInfo.Samples = append(segInfo.Samples, &SampleInfo{dur, fragIdx, pts})
